requirements.txt
certifi==2020.12.5
chardet==4.0.0
fastapi==0.63.0
idna==2.10
Pillow==8.2.0
pydantic==1.8.1
requests==2.25.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
